+++ blocklets/ai-studio/api/src/routes/dataset/documents.ts
@@ -13,16 +13,16 @@ import { userAuth } from '../../libs/user';
 import DatasetContent from '../../store/models/dataset/content';
 import Dataset from '../../store/models/dataset/dataset';
 import DatasetDocument from '../../store/models/dataset/document';
-import DatasetSegment from '../../store/models/dataset/segment';
-import VectorStore from '../../store/vector-store';
-import { discussionsIterator, queue } from './embeddings';
+import EmbeddingHistories from '../../store/models/dataset/embedding-history';
+import FaissStore from '../../store/vector-store-faiss';
+import { getDiscussionIds, queue, updateHistoriesAndStore } from './embeddings';
 
 const router = Router();
 const upload = multer({ storage: multer.memoryStorage() });
 
 export interface CreateDiscussionItem {
   name: string;
-  data: { type: 'discussion'; fullSite?: boolean; id: string };
+  data: { type: 'discussion'; fullSite?: boolean; types?: ('discussion' | 'blog' | 'doc')[]; id: string };
 }
 
 export type CreateDiscussionItemInput = CreateDiscussionItem | CreateDiscussionItem[];
@@ -80,10 +80,25 @@ router.get('/:datasetId/search', async (req, res) => {
   }
 
   const embeddings = new AIKitEmbeddings({});
-  const store = await VectorStore.load(datasetId, embeddings);
-  const docs = await store.similaritySearch(input.message, 4);
+  const store = await FaissStore.load(datasetId, embeddings);
 
-  res.json({ docs: docs.map((i) => i.pageContent) });
+  try {
+    // const records = await UpdateHistories.findAll({ where: { datasetId }, attributes: ['segmentId'] });
+    // const uniqueSegmentIds = [...new Set(records.map((record) => record.segmentId).flat())];
+
+    if (store.getMapping() && !Object.keys(store.getMapping()).length) {
+      res.json({ docs: [] });
+      return;
+    }
+
+    const docs = await store.similaritySearch(input.message, 3);
+    const result = docs.map((x) => x?.pageContent);
+
+    res.json({ docs: result });
+  } catch (error) {
+    console.error(error);
+    res.json({ docs: [] });
+  }
 });
 
 /**
@@ -176,10 +191,12 @@ router.delete('/:datasetId/documents/:documentId', user(), userAuth(), async (re
 
   const document = DatasetDocument.findOne({ where: { id: documentId, datasetId } });
 
+  await updateHistoriesAndStore(datasetId, documentId);
+
   await Promise.all([
     DatasetDocument.destroy({ where: { id: documentId, datasetId } }),
-    DatasetSegment.destroy({ where: { documentId } }),
     DatasetContent.destroy({ where: { documentId } }),
+    EmbeddingHistories.destroy({ where: { documentId, datasetId } }),
   ]);
 
   res.json(document);
@@ -241,15 +258,27 @@ router.post('/:datasetId/documents/text', user(), userAuth(), async (req, res) =
 router.post('/:datasetId/documents/discussion', user(), async (req, res) => {
   const { datasetId } = req.params;
 
-  const createItemsSchema = Joi.object<CreateDiscussionItem>({
+  const createItemsSchema = Joi.object({
     name: Joi.string().empty(['', null]),
     data: Joi.object({
       type: Joi.string().valid('discussion').required(),
       fullSite: Joi.boolean(),
+      types: Joi.array().items(Joi.string()).sparse(false),
       id: Joi.string().empty(['', null]),
-    }).required(),
+    })
+      .required()
+      .when('.fullSite', {
+        is: true,
+        then: Joi.object({
+          types: Joi.array().items(Joi.string()).min(1).required(),
+          id: Joi.any().optional(),
+        }),
+        otherwise: Joi.object({
+          types: Joi.array().items(Joi.string()).sparse(true).default([]),
+          id: Joi.string().required(),
+        }),
+      }),
   });
-
   const createItemInputSchema = Joi.alternatives<CreateDiscussionItemInput>().try(
     Joi.array().items(createItemsSchema),
     createItemsSchema
@@ -283,20 +312,14 @@ router.post('/:datasetId/documents/discussion', user(), async (req, res) => {
   let docs: DatasetDocument[] = [];
   const fullSite = arr.find((x) => x.data?.fullSite);
   if (fullSite) {
-    const ids = [];
-    for await (const { id: discussionId } of discussionsIterator()) {
-      try {
-        ids.push(discussionId);
-      } catch (error) {
-        console.error(`embedding discussion ${discussionId} error`, { error });
-      }
-    }
+    const ids = await getDiscussionIds(fullSite.data.types || []);
 
     const document = await DatasetDocument.create({
       type: 'fullSite',
       data: {
         type: 'fullSite',
         ids,
+        types: fullSite.data.types || [],
       },
       name: fullSite.name,
       datasetId,
@@ -399,9 +422,7 @@ router.put('/:datasetId/documents/:documentId/text', user(), userAuth(), async (
 
   const document = await DatasetDocument.findOne({ where: { id: documentId, datasetId } });
 
-  if (document) {
-    queue.push({ documentId: document.id });
-  }
+  if (document) queue.push({ documentId: document.id });
 
   res.json(document);
 });
@@ -464,9 +485,7 @@ router.put('/:datasetId/documents/:documentId/file', user(), userAuth(), upload.
 
   const document = await DatasetDocument.findOne({ where: { id: documentId, datasetId } });
 
-  if (document) {
-    queue.push({ documentId: document.id });
-  }
+  if (document) queue.push({ documentId: document.id });
 
   res.json(document);
 });
@@ -498,9 +517,7 @@ router.post('/:datasetId/documents/:documentId/embedding', user(), userAuth(), a
 
   const [document] = await Promise.all([DatasetDocument.findOne({ where: { id: documentId } })]);
 
-  if (document) {
-    queue.push({ documentId: document.id });
-  }
+  if (document) queue.push({ documentId: document.id });
 
   res.json(document);
 });